# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Money Management API\uff08\u30bb\u30ad\u30e5\u30ea\u30c6\u30a3\u5f37\u5316\u7248\uff09",
...
        "version": "2.0.0",
        "contact": {
            "name": "Money Management Team",
            "email": "support@moneymanagement.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "\u958b\u767a\u74b0\u5883"
        },
        {
            "url": "https://localhost:443",
            "description": "HTTPS\u74b0\u5883\uff08\u672c\u756a\u5bfe\u5fdc\uff09"
        }
    ],
    "security": [
        {
            "bearerAuth": [],
            "csrfToken": []
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "summary": "\u65b0\u898f\u30e6\u30fc\u30b6\u30fc\u767b\u9332",
                "tags": [
                    "\u8a8d\u8a3c"
                ],
                "security": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "name",
                                    "account_id",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "\u5c71\u7530\u592a\u90ce"
                                    },
                                    "account_id": {
                                        "type": "string",
                                        "example": "yamada_taro"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "\u767b\u9332\u6210\u529f",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "\u30ea\u30af\u30a8\u30b9\u30c8\u30a8\u30e9\u30fc",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "\u30ec\u30fc\u30c8\u5236\u9650\u8d85\u904e",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "summary": "\u30ed\u30b0\u30a4\u30f3",
                "tags": [
                    "\u8a8d\u8a3c"
                ],
                "security": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "account_id",
                                    "password"
                                ],
                                "properties": {
                                    "account_id": {
                                        "type": "string",
                                        "example": "yamada_taro"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "\u30ed\u30b0\u30a4\u30f3\u6210\u529f",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\u8a8d\u8a3c\u5931\u6557",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "\u30ec\u30fc\u30c8\u5236\u9650\u8d85\u904e",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/me": {
            "get": {
                "summary": "\u73fe\u5728\u30e6\u30fc\u30b6\u30fc\u60c5\u5831\u53d6\u5f97",
                "tags": [
                    "\u8a8d\u8a3c"
                ],
                "responses": {
                    "200": {
                        "description": "\u30e6\u30fc\u30b6\u30fc\u60c5\u5831",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\u8a8d\u8a3c\u30a8\u30e9\u30fc",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "summary": "\u30ed\u30b0\u30a2\u30a6\u30c8\uff08\u73fe\u5728\u30c8\u30fc\u30af\u30f3\u7121\u52b9\u5316\uff09",
                "tags": [
                    "\u8a8d\u8a3c"
                ],
                "responses": {
                    "200": {
                        "description": "\u30ed\u30b0\u30a2\u30a6\u30c8\u6210\u529f",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "\u30ed\u30b0\u30a2\u30a6\u30c8\u3057\u307e\u3057\u305f"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "\u8a8d\u8a3c\u30a8\u30e9\u30fc",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "csrfToken": {
                "type": "apiKey",
                "in": "header",
                "name": "X-CSRF-Token"
            }
        },
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "account_id": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "AuthResponse": {
                "type": "object",
                "properties": {
                    "token": {
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                }
            },
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string"
                    }
                }
            }
        }
    },
...
}